moving validation code into validator
